Move serverless API functions to root
